Refactoring api calls and moving them to a service
