+++ types/ApplicantTokenPair.ts
@@ -1,0 +1,6 @@
+interface ApplicantTokenPair {
+  applicantId: string;
+  sdkToken: string;
+}
+
+export default ApplicantTokenPair;
